added settings page and file explorer

+++ lib/Apps/terminal/commands/bash_commands/make_dir.dart
@@ -32,8 +32,7 @@ class Mkdir implements DecodeCommand {
     if (error.isEmpty) {
       Shell shell = Shell.init();
       shell.create(path + "/$fileName");
-      Directory newDir = Directory(path + "/$fileName");
-      fileController.add(newDir);
+      fileController.updateUi(path + "/$fileName");
       return "";
     } else
       return error;
@@ -42,33 +41,54 @@ class Mkdir implements DecodeCommand {
 
 class Rmdir implements DecodeCommand {
   @override
-  dynamic executeCommand(BuildContext context, int id, String fileName) {
+  dynamic executeCommand(BuildContext context, int id, String command) {
     final controller = Get.find<TerminalController>();
     final fileController = Get.find<FileController>();
-    return rmdir(fileController, controller.path, fileName);
+    String path, fileName;
+    //compiling the path
+    if (command.startsWith("/") || command.split("/").length >= 2) {
+      var a = command.split("/");
+      a.removeLast();
+      path = controller.path + "/" + a.join("/").replaceFirst("/", "");
+      fileName = command.split("/").last;
+    } else if (command.isNotEmpty) {
+      var a = command.split("/");
+      a.removeLast();
+      path = controller.path + "/" + a.join("/");
+      fileName = command;
+    } else
+      fileName = "";
+    return rmdir(fileController, path, fileName);
   }
 
   rmdir(FileController fileController, String path, String fileName) {
-    print("Rm ing");
     Ls ls = Ls();
     List items = ls.ls(path);
+    String error = "";
 
-    String error = "File Not Found \n\nNavigate to the Working Directory";
-    for (var item in items) {
-      if (item is Directory &&
-          item.path.split("/").last.trim() == fileName.trim()) {
-        error = "";
-        break;
+    if (fileName.isEmpty)
+      error = "Specify a name";
+    else {
+      for (var item in items) {
+        List split = item.path.split("/");
+        split.removeLast();
+        if (item is Directory &&
+            item.path.trim() == split.join("/") + "/$fileName") {
+          break;
+        }
       }
     }
-    items = ls.ls(path + "/$fileName");
-    if (items.isNotEmpty) error = "Directory is not empty";
+
     if (error.isEmpty) {
-      Shell shell = Shell.init();
-      shell.remove(path + "/$fileName");
-      print("renoved");
-      fileController.delete(null);
-      return "";
+      var dirs = ls.ls(path +
+          "/$fileName"); //checking if the selected folder is empty or not
+      if (dirs.isEmpty) {
+        Shell shell = Shell.init();
+        shell.remove(path + "/$fileName");
+        fileController.updateUi(path + "/$fileName");
+        return "";
+      } else
+        error = "Directory is not empty";
     }
 
     return error;
